+++ lib/src/gherkin/attachments/attachment.dart
@@ -1,0 +1,7 @@
+class Attachment {
+  final String data;
+  final String mimeType;
+  final String context;
+
+  Attachment(this.data, this.mimeType, [this.context]);
+}