+++ lib/src/reporters/json/json_embedding.dart
@@ -1,0 +1,11 @@
+class JsonEmbedding {
+  String mimeType;
+  String data;
+
+  Map<String, dynamic> toJson() {
+    return {
+      'mime_type': mimeType,
+      'data': data,
+    };
+  }
+}